chore: rename package to graphback-core
