+++ packages/graphback-codegen-resolvers/src/layered/LayeredResolverGen.ts
@@ -1,8 +1,8 @@
-import { InputModelTypeContext } from "@graphback/codegen-core"
+import { InputModelTypeContext } from "@graphback/core"
 
 /**
  * Generate source code for resolvers layer using Apollo GraphQL format
- * and injected service layer. Service layer offers various capabilities like monitoring, cache etc. 
+ * and injected service layer. Service layer offers various capabilities like monitoring, cache etc.
  * so resolver logic can be kept simple and interchangable.
  */
 export class LayeredResolverGen {
@@ -17,7 +17,7 @@ export class LayeredResolverGen {
   public generate() {
     const resolvers = {};
     for (const resolverElement of this.inputContext) {
-      //TODO 
+      //TODO
     }
 
     return resolvers;
