+++ packages/graphback-core/tests/mock.graphql
