+++ packages/graphback-db-manage/src/database/initialization/DatabaseInitializationStrategy.ts
@@ -1,4 +1,4 @@
-import { InputModelTypeContext } from '@graphback/codegen-core';
+import { InputModelTypeContext } from '@graphback/core';
 import { Change } from '@graphql-inspector/core';
 import { DatabaseContextProvider } from '../migrations/DatabaseContextProvider';
