+++ packages/graphback-runtime/src/service/GraphbackCRUDService.ts
@@ -1,26 +1,26 @@
-import { InputModelTypeContext } from "@graphback/codegen-core"
+import { InputModelTypeContext } from "@graphback/core"
 import { AdvancedFilter } from '../data/GraphbackDataProvider';
 
 
 /**
- * Graphback layered architecture component that can be called 
+ * Graphback layered architecture component that can be called
  * from the resolver layer in GraphQL and Middlerware layer in RESTfull approach.
- * 
+ *
  * Graphback implements server side procesing using following flow:
- * 
- * `GraphQL Resolvers` ->  `GraphbackCRUDService` [1-*] -> `GraphbackDataProvider` 
- * 
+ *
+ * `GraphQL Resolvers` ->  `GraphbackCRUDService` [1-*] -> `GraphbackDataProvider`
+ *
  * Services can be composable (each service can reference multiple layers of other services).
- * For data abstraction Graphback `GraphbackDataProvider` can be being used. 
- * 
+ * For data abstraction Graphback `GraphbackDataProvider` can be being used.
+ *
  * @see GraphbackDataProvider
  */
 // tslint:disable-next-line: no-any
 export interface GraphbackCRUDService<Type = any, GraphbackContext = any> {
 
     /**
-     * Implementation for object creation 
-     * 
+     * Implementation for object creation
+     *
      * @param name name of the object to create
      * @param data input data
      * @param context context object passed from graphql or rest layer
@@ -28,8 +28,8 @@ export interface GraphbackCRUDService<Type = any, GraphbackContext = any> {
     create(inputType: InputModelTypeContext, data: Type, context?: GraphbackContext): Promise<Type>;
 
     /**
-     * Implementation for object updates 
-     * 
+     * Implementation for object updates
+     *
      * @param name name of the object to create
      * @param id of the object to update
      * @param data input data
@@ -38,17 +38,17 @@ export interface GraphbackCRUDService<Type = any, GraphbackContext = any> {
     update(inputType: InputModelTypeContext, id: string, data: Type, context?: GraphbackContext): Promise<Type>;
 
     /**
-     * Implementation for object deletes 
-     * 
+     * Implementation for object deletes
+     *
      * @param name name of the object to create
-     * @param data input data containing id 
+     * @param data input data containing id
      * @param context context object passed from graphql or rest layer
      */
     delete(inputType: InputModelTypeContext, id: string, context?: GraphbackContext): Promise<string>;
 
     /**
-     * Implementation for reading object 
-     * 
+     * Implementation for reading object
+     *
      * @param name name of the object to create
      * @param id id of the object
      * @param context context object passed from graphql or rest layer
@@ -57,7 +57,7 @@ export interface GraphbackCRUDService<Type = any, GraphbackContext = any> {
 
     /**
      * Implementation for finding all objects
-     * 
+     *
      * @param name name of the object to create
      * @param id id of the object
      * @param context context object passed from graphql or rest layer
@@ -66,7 +66,7 @@ export interface GraphbackCRUDService<Type = any, GraphbackContext = any> {
 
     /**
      * Implementation for reading objects with filtering capabilities
-     * 
+     *
      * @param name name of the object to create
      * @param filter filter by specific type
      * @param context context object passed from graphql or rest layer
@@ -75,7 +75,7 @@ export interface GraphbackCRUDService<Type = any, GraphbackContext = any> {
 
     /**
      * Subscription for all creation events
-     * 
+     *
      * @param name name of the component to subscribe
      * @param context additional context
      */
@@ -83,7 +83,7 @@ export interface GraphbackCRUDService<Type = any, GraphbackContext = any> {
 
     /**
      * Subscription for all update events
-     * 
+     *
      * @param name name of the component to subscribe
      * @param context additional context
      */
@@ -91,7 +91,7 @@ export interface GraphbackCRUDService<Type = any, GraphbackContext = any> {
 
     /**
      * Subscription for all deletion events
-     * 
+     *
      * @param name name of the component to subscribe
      * @param context additional context
      */
