+++ packages/graphback/package.json
@@ -47,7 +47,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@graphback/codegen-schema": "0.10.0-dev1",
-    "@graphback/codegen-core": "0.10.0-dev1",
+    "@graphback/core": "0.10.0-dev1",
     "@graphback/codegen-client": "0.10.0-dev1",
     "@graphback/codegen-resolvers": "0.10.0-dev1",
     "@graphback/runtime": "0.10.0-dev1",
