+++ packages/graphback/src/index.ts
@@ -6,7 +6,7 @@ export * from './IGraphQLBackend'
 
 export * from '@graphback/codegen-schema'
 export * from '@graphback/codegen-client'
-export * from '@graphback/codegen-core'
+export * from '@graphback/core'
 export * from '@graphback/codegen-resolvers'
 export * from '@graphback/runtime'
 export * from '@graphback/db-manage';
